Adds mortars (#223)
* Ports RMC mortar files and gets them into a half functional state

At the moment the game can build. The mortar and its ammo exist ingame and the mortar can be deployed. Trying to open the UI will call the open function but the UI does not actually open. Explosions are also entirely commented out and the code to support it have not been ported. Locale files are also not ported. All mortar related sprites have been ported.

* Cleans the UI and unnecessary RMC dependencies

The blue horizontal line was removed as it would require editing base ss14 UI functions which I dont want to fuck with. The camera button on the mortar has been removed and the max dial section was moved above the dial buttons. RMC pop out window related code was removed.

* Comments out a debug assert which should not exist

Goes off because nomad is the name of the fallback job instead of passenger for some reason. Honestly no idea why there is an assert for this.

* Ports the locale file

A few things were removed too such as the ones about the mortar camera and the planet.

* Ports mortar sfx, fixes mortar animation and cleans some code

* Setup mortar sprites and fixed some other issues

New inhands, undeployed and firing sprites my me via editing civ13 mortar sprites. All the sfx are the same as the RMC ones besides the melee hit sfx.

* "old mortar" added

Uses some terrible programmer sprites for the inhands which should honestly be redone.

* Adds navigation tools for aiming the mortars

They all just function like a GPS for now and tell your current location x and y.

* Tweaks some values and disables warning messages

* Takes out the unused RMC sprites

I forgot to take these out after i replaced them