Sound updates 1: Explosions and Mortars (#233)
* Ports a ton of civ13 sounds and some explosion sound collections

* Sets up custom explosions

Needs lots of work still. Also removed some bad sfx and renamed some

* Some more yml changes

* Changes volume of explosion files

I really wish this could be done in the YML but I cant figure out how to get it working so im moving on before I waste more time on this. The volume is adjusted in by changing the volume of the file its self.

* Deletes dupe sfx

* Mortar sfx

Implements civ13 sfx for firing mortar shells and landing shells

* Fix some file names part 1

* rename some files part 2

has to be done in two parts because the windows file system is stupid like that